+++ test/NCalc.Tests/Fixtures/FactoriesFixtures.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using NCalc.Antlr.Configuration;
 using NCalc.DependencyInjection.Configuration;
 using NCalc.Factories;
 using NCalc.MemoryCache.Configuration;
@@ -18,6 +19,19 @@ public sealed class FactoriesFixture : FactoriesFixtureBase
     {
         var serviceProvider = new ServiceCollection()
             .AddNCalc()
+            .Services.BuildServiceProvider();
+        ExpressionFactory = serviceProvider.GetRequiredService<IExpressionFactory>();
+        LogicalExpressionFactory = serviceProvider.GetRequiredService<ILogicalExpressionFactory>();
+    }
+}
+
+public sealed class FactoriesWithAntlrFixture : FactoriesFixtureBase
+{
+    public FactoriesWithAntlrFixture() 
+    {
+        var serviceProvider = new ServiceCollection()
+            .AddNCalc()
+            .WithAntlr()
             .Services.BuildServiceProvider();
         ExpressionFactory = serviceProvider.GetRequiredService<IExpressionFactory>();
         LogicalExpressionFactory = serviceProvider.GetRequiredService<ILogicalExpressionFactory>();